Spring

+++ src/main/java/pl/training/shop/commons/ResultPage.java
@@ -9,7 +9,7 @@ public class ResultPage<T> {
 
     List<T> data;
     int pageNumber;
-    int totalPages;
+    long totalPages;
 
     public <OT> ResultPage<OT> with(List<OT> data) {
         return new ResultPage<>(data, pageNumber, totalPages);
